refactor: moves Notifications out of a folder
